Moved common integration stuff to a file
